abstracting out the dll parser into its own package to be shared
