+++ openvmm/openvmm_pcat_locator/src/lib.rs
@@ -33,17 +33,6 @@ fn system32_path() -> String {
     format!("{windows_dir}\\System32")
 }
 
-/// expose function for parsing a dll resource
-pub fn try_find_resource_from_dll(
-    file: &fs_err::File,
-    resource_type: &[u8; 4],
-    resource_code: u32,
-) -> anyhow::Result<Option<(u64, usize)>> {
-    // Create a descriptor for the VMFW resource with the given ID
-    let descriptor = DllResourceDescriptor::new(resource_type, resource_code);
-    resource_dll_parser::try_find_resource_from_dll(file, &descriptor)
-}
-
 /// Attempt to automatically find and open the PCAT BIOS. Will always prefer
 /// the more recently updated vmfirmwarepcat.dll over vmfirmware.dll.
 pub fn find_pcat_bios(command_line_path: Option<&Path>) -> anyhow::Result<RomFileLocation> {
@@ -161,29 +150,4 @@ fn parse_rom_file(
         start,
         len,
     })
-}
-
-pub(crate) struct DllResourceDescriptor {
-    /// 4 characters encoded in LE UTF-16
-    resource_type: [u8; 8],
-    id: u32,
-}
-
-impl DllResourceDescriptor {
-    const fn new(resource_type: &[u8; 4], id: u32) -> Self {
-        Self {
-            id,
-            // Convert to LE UTF-16, only support ASCII names today
-            resource_type: [
-                resource_type[0],
-                0,
-                resource_type[1],
-                0,
-                resource_type[2],
-                0,
-                resource_type[3],
-                0,
-            ],
-        }
-    }
 }
