+++ vm/vmgs/vmgstool/Cargo.toml
@@ -22,7 +22,6 @@ disk_vhd1.workspace = true
 uefi_nvram_storage.workspace = true
 guid.workspace = true
 hcl_compat_uefi_nvram_storage.workspace = true
-openvmm_pcat_locator.workspace = true
 pal_async.workspace = true
 uefi_nvram_specvars.workspace = true
 uefi_specs.workspace = true
@@ -39,6 +38,7 @@ serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true, features = ["std"] }
 thiserror.workspace = true
 ucs2.workspace = true
+resource_dll_parser.workspace = true
 
 [dev-dependencies]
 tempfile.workspace = true
